requirements.txt
matplotlib==3.7.1
pandas==1.5.3
Pillow==9.4.0
Pillow==10.0.1
Pillow==10.1.0
seaborn==0.13.0
streamlit==1.28.2
